requirements.txt
websockets>=12.0
aiohttp>=3.9
aiofiles>=23.0
google-auth>=2.28
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
